// src/types.ts
/**
 * @license
 * SPDX-License-Identifier: Apache-2.0
 */

export interface Point {
  x: number;
  y: number;
}

export type Direction = 'UP' | 'DOWN' | 'LEFT' | 'RIGHT';

export interface Track {
  id: string;
  title: string;
  artist: string;
  duration: string;
  color: string;
}

export interface GameState {
  snake: Point[];
  food: Point;
  direction: Direction;
  isPaused: boolean;
  isGameOver: boolean;
  score: number;
  highScore: number;
}